refactor: split tax advantaged limit meters
