+++ frontend/src/accounts/AccountsPage.tsx
@@ -8,7 +8,7 @@ import AccountListSection from '@/accounts/components/AccountListSection'
 import AccountSummaryStatement from '@/accounts/components/AccountSummaryStatement'
 import AccountsMetricsBand from '@/accounts/components/AccountsMetricsBand'
 import ArchivedAccountsSection from '@/accounts/components/ArchivedAccountsSection'
-import TaxAdvantagedLimitsSection from '@/accounts/components/TaxAdvantagedLimitsSection'
+import TaxAdvantagedLimitsSection from '@/accounts/components/tax-advantaged-limits/TaxAdvantagedLimitsSection'
 import { useAccountFilters } from '@/accounts/hooks/useAccountFilters'
 import { useAccountSections } from '@/accounts/hooks/useAccountSections'
 import { useAccountsMetrics } from '@/accounts/hooks/useAccountsMetrics'
